v5.0.0 changes for pure11

+++ src/Data/Number/Format.cc
@@ -18,7 +18,7 @@
 namespace Data_Number_Format {
   using namespace PureScript;
   std::string to_string(const any& f) {
-    std::string str = std::to_string (f);
+    std::string str = std::to_string ((double)f);
     int offset{1}; if (str.find_last_not_of('0') == str.find('.')) { offset = 0; }
     str.erase(str.find_last_not_of('0') + offset, std::string::npos);
     return str;
